(chore) clinical dashboard cleanups (#1707)
* (chore) clinical dashboard cleanups

* removes log residue

* build fixes with cacx

* refreshes cacx-visits folder
