+++ .gitignore
@@ -74,3 +74,6 @@ typings/
 .next/**
 build/**
 **/*dist
+
+# vscode
+.vscode
