+++ frontend/spa-build-config.json
@@ -1,11 +1,12 @@
 {
   "frontendModules": {
-    "@ohri/openmrs-esm-ohri-cervical-cancer-app": "next",
     "@ohri/openmrs-esm-ohri-core-app": "next",
-    "@ohri/openmrs-esm-ohri-covid-app": "next",
     "@ohri/openmrs-esm-ohri-form-render-app": "next",
     "@ohri/openmrs-esm-ohri-hiv-app": "next",
+    "@ohri/openmrs-esm-ohri-covid-app": "next",
     "@ohri/openmrs-esm-ohri-pmtct": "next",
+    "@ohri/openmrs-esm-ohri-tb-app": "next",
+    "@ohri/openmrs-esm-ohri-cervical-cancer-app": "next",
     "@openmrs/esm-appointments-app": "next",
     "@openmrs/esm-dispensing-app": "next",
     "@openmrs/esm-form-builder-app": "next",
@@ -30,7 +31,18 @@
     "@openmrs/esm-patient-vitals-app": "next",
     "@openmrs/esm-primary-navigation-app": "next",
     "@openmrs/esm-patient-registration-app": "next",
-    "@openmrs/esm-devtools-app": "next"
+    "@openmrs/esm-devtools-app": "next",
+    "@openmrs/esm-openconceptlab-app": "next",
+    "@openmrs/esm-generic-patient-widgets-app": "next",
+    "@openmrs/esm-patient-lists-app": "next",
+    "@openmrs/esm-cohort-builder-app": "next",
+    "@openmrs/esm-patient-list-management-app": "next",
+    "@openmrs/esm-service-queues-app": "next",
+    "@openmrs/esm-form-entry-app": "next",
+    "@openmrs/esm-patient-labs-app": "next",
+    "@openmrs/esm-fast-data-entry-app": "next",
+    "@openmrs/esm-patient-orders-app": "next",
+    "@openmrs/esm-system-admin-app": "next"
   },
   "spaPath": "$SPA_PATH",
   "apiUrl": "$API_URL",
