+++ packages/esm-cervical-cancer-app/src/index.ts
@@ -1,6 +1,8 @@
-import { defineConfigSchema, getAsyncLifecycle, getSyncLifecycle, provide } from '@openmrs/esm-framework';
+import { defineConfigSchema, getAsyncLifecycle, getSyncLifecycle } from '@openmrs/esm-framework';
 import { caCxSummaryDashboardMeta, caCxVisitsDashboardMeta, cervicalCancerFolderMeta } from './dashboard.meta';
 import { createDashboardGroup, createDashboardLink } from '@openmrs/esm-patient-common-lib';
+import CacxSummaryList from './views/cacx-summary/cacx-summary.component';
+import CacxVisitServices from './views/cacx-visits/cacx-visits-services.component';
 
 export const importTranslation = require.context('../translations', false, /.json$/, 'lazy');
 
@@ -23,7 +25,7 @@ export const cacxSummaryDashboardLink = getSyncLifecycle(
   createDashboardLink({ ...caCxSummaryDashboardMeta, moduleName }),
   options,
 );
-export const cacxSummaryDashboard = getAsyncLifecycle(() => import('./views/cacx-summary/cacx-summary.component'), {
+export const cacxSummaryDashboard = getSyncLifecycle(CacxSummaryList, {
   featureName: 'cacx-summary',
   moduleName,
 });
@@ -32,10 +34,7 @@ export const cacxVisitsDashboardLink = getSyncLifecycle(
   createDashboardLink({ ...caCxVisitsDashboardMeta, moduleName }),
   options,
 );
-export const cacxVisitsDashboard = getAsyncLifecycle(
-  () => import('./views/cacx-visits/cacx-visits-services.component'),
-  {
-    featureName: 'cacx-visits',
-    moduleName,
-  },
-);
+export const cacxVisitsDashboard = getSyncLifecycle(CacxVisitServices, {
+  featureName: 'cacx-visits',
+  moduleName,
+});
