+++ packages/esm-form-render-app/src/constants.ts
@@ -2,8 +2,3 @@ export const HTSEncounterType = '30b849bd-c4f4-4254-a033-fe9cf01001d8';
 export const ConceptTrue = 'cf82933b-3f3f-45e7-a5ab-5d31aaee3da3';
 export const ConceptFalse = '488b58ff-64f5-4f8a-8979-fa79940b1594';
 export const UnspecifiedValue = 'OHRI_UNSPECIFIED';
-export const encounterRepresentation =
-  'custom:(uuid,encounterDatetime,encounterType,location:(uuid,name),' +
-  'patient:(uuid,display),encounterProviders:(uuid,provider:(uuid,name)),' +
-  'obs:(uuid,obsDatetime,voided,groupMembers,concept:(uuid,name:(uuid,name)),value:(uuid,name:(uuid,name),' +
-  'names:(uuid,conceptNameType,name))))';
