+++ packages/esm-hiv-app/src/views/hts/care-and-treatment/summary-tiles/ct-summary-tiles.component.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { OHRIProgrammeSummaryTiles, getReportingCohort } from '@ohri/openmrs-esm-ohri-commons-lib';
-import { clientsEnrolledToCare } from '../../../../../constants';
+import { clientsEnrolledToCare } from '../../../../constants';
 
 function CTSummaryTiles({ launchWorkSpace }) {
   const { t } = useTranslation();
